Unstable but working ads
The idea of the ads is complete. All kind of ads show up and close with succes.

Editor and runtime on phone (android) do have different results. With reward click the editor will get the rewards and the UI will be updated but on the phone it won't update unless I manually call the function to do that. I have to look into that.

Nonetheless, the ads work which was a thing I wanted to learn.